move datamodel to fixtures
