+++ packages/record/src/redwoodrecord/__tests__/RedwoodRecord.test.js
@@ -1,10 +1,9 @@
 import * as ValidationErrors from '@redwoodjs/api'
 
+import datamodel from '../__fixtures__/datamodel.js'
 import RedwoodRecord from '../RedwoodRecord'
 import Reflection from '../Reflection'
 import RelationProxy from '../RelationProxy'
-
-import datamodel from './datamodel.js'
 
 const db = { user: jest.fn() }
